moved politico into scrapers file
